Fix compile warning

+++ Shared/EnHeaderCtrl.cpp
@@ -906,7 +906,7 @@ int CEnHeaderCtrl::OnToolHitTest(CPoint point, TOOLINFO* pTI) const
 	if ((nItem != -1) && IsItemVisible(nItem))
 	{
 #ifdef _DEBUG
-		if (m_tooltips.GetLastHitToolInfo().uId != nItem)
+		if (m_tooltips.GetLastHitToolInfo().uId != (UINT)nItem)
 			TRACE(_T("CEnHeaderCtrl::OnToolHitTest(%d, %s)\n"), nItem, GetItemToolTip(nItem));
 #endif
